+++ packages/nhsNotifyUpdateCallback/src/lambdaHandler.ts
@@ -9,69 +9,10 @@ import httpHeaderNormalizer from "@middy/http-header-normalizer"
 
 import errorHandler from "@nhs/fhir-middy-error-handler"
 
-import {createHmac, timingSafeEqual} from "crypto"
 import {MessageStatusResponse} from "./types"
+import {checkSignature, response} from "./helpers"
 
 export const logger = new Logger({serviceName: "nhsNotifyUpdateCallback"})
-
-const APP_NAME = process.env.APP_NAME ?? "NO-APP-NAME"
-const API_KEY = process.env.API_KEY ?? "NO-API-KEY"
-
-function response(statusCode: number, body: unknown = {}) {
-  return {
-    statusCode,
-    body: JSON.stringify(body)
-  }
-}
-
-/**
- * Checks the incoming NHS Notify request signature.
- * If it's okay, returns undefined.
- * If it's not okay, it returns the error response object.
- */
-function checkSignature(event: APIGatewayProxyEvent) {
-  const signature = event.headers["x-hmac-sha256-signature"]
-  if (!signature) {
-    logger.error("No x-hmac-sha256-signature header given")
-    return response(401, {message: "No x-hmac-sha256-signature given"})
-  }
-
-  const givenApiKey = event.headers["x-api-key"]
-  if (!givenApiKey) {
-    logger.error("No x-api-key header given")
-    return response(401, {message: "No x-api-key header given"})
-  }
-
-  // Compute the HMAC-SHA256 hash of the combination of the request body and the secret value
-  const secretValue = `${APP_NAME}.${API_KEY}`
-  const payload = event.body ?? ""
-
-  // Compute the HMAC as a Buffer
-  const expectedSigBuf = createHmac("sha256", secretValue)
-    .update(payload, "utf8")
-    .digest() // Buffer
-
-  // Convert the incoming hex signature into a Buffer
-  let givenSigBuf: Buffer
-  try {
-    givenSigBuf = Buffer.from(signature, "hex")
-  } catch {
-    logger.error("Invalid hex in signature header", {givenSignature: signature})
-    return response(403, {message: "Malformed signature"})
-  }
-
-  // Must be same length for timingSafeEqual
-  if (givenSigBuf.length !== expectedSigBuf.length ||
-      !timingSafeEqual(expectedSigBuf, givenSigBuf)) {
-    logger.error("Incorrect signature given", {
-      expectedSignature: expectedSigBuf.toString("hex"),
-      givenSignature: signature
-    })
-    return response(403, {message: "Incorrect signature"})
-  }
-
-  return undefined
-}
 
 const lambdaHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   logger.appendKeys({
@@ -85,9 +26,10 @@ const lambdaHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPro
   if (!event.headers["x-request-id"]) return response(401, {message: "No x-request-id given"})
 
   // Check the request signature
-  const isErr = checkSignature(event)
+  const isErr = checkSignature(logger, event)
   if (isErr) return isErr
 
+  // Parse out the request body
   if (!event.body) return response(401, {message: "No request body given"})
   try {
     const payload: MessageStatusResponse = JSON.parse(event.body)
@@ -96,6 +38,7 @@ const lambdaHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPro
     logger.error("Failed to parse payload", {error, payload: event.body})
   }
 
+  // All's well that ends well
   return {
     statusCode: 202,
     body: "OK"
